train model to predict t+1 internet traffic data values
These changes add capabilities to train and save a model to predict t+1 internet traffic data values.

[#2]
