+++ anomaly-detection-train/src/main/java/com/orange/cloud/anomalydetection/train/AnomalyDetectionTrainApplication.java
@@ -1,0 +1,13 @@
+package com.orange.cloud.anomalydetection.train;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AnomalyDetectionTrainApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(AnomalyDetectionTrainApplication.class, args);
+    }
+
+}
