+++ anomaly-detection-train/src/main/java/com/orange/cloud/anomalydetection/train/TrainingServiceDefault.java
@@ -1,0 +1,101 @@
+package com.orange.cloud.anomalydetection.train;
+
+import org.deeplearning4j.api.storage.StatsStorage;
+import org.deeplearning4j.eval.RegressionEvaluation;
+import org.deeplearning4j.nn.api.OptimizationAlgorithm;
+import org.deeplearning4j.nn.conf.MultiLayerConfiguration;
+import org.deeplearning4j.nn.conf.NeuralNetConfiguration;
+import org.deeplearning4j.nn.conf.Updater;
+import org.deeplearning4j.nn.conf.layers.GravesLSTM;
+import org.deeplearning4j.nn.conf.layers.RnnOutputLayer;
+import org.deeplearning4j.nn.multilayer.MultiLayerNetwork;
+import org.deeplearning4j.nn.weights.WeightInit;
+import org.deeplearning4j.optimize.listeners.ScoreIterationListener;
+import org.deeplearning4j.ui.api.UIServer;
+import org.deeplearning4j.ui.stats.StatsListener;
+import org.deeplearning4j.ui.storage.InMemoryStatsStorage;
+import org.nd4j.linalg.activations.Activation;
+import org.nd4j.linalg.api.ndarray.INDArray;
+import org.nd4j.linalg.dataset.DataSet;
+import org.nd4j.linalg.lossfunctions.LossFunctions;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Component;
+
+/**
+ * This example was inspired by DL4j regression examples, found here:
+ * https://github.com/deeplearning4j/dl4j-examples/blob/master/dl4j-examples/src/main/java/org/deeplearning4j/examples/recurrent/regression/SingleTimestepRegressionExample.java
+ *
+ * It demonstrates single time step regression using LSTM
+ */
+@Component
+public class TrainingServiceDefault implements TrainingService {
+
+    public static final int N_EPOCHS = 100;
+    public static final double LEARNING_RATE = 0.00005;
+    private static final Logger LOGGER = LoggerFactory.getLogger(TrainingServiceDefault.class);
+
+    @Override
+    public MultiLayerNetwork train(DataSet trainData, DataSet testData) {
+
+        // ----- Configure the network -----
+        MultiLayerConfiguration conf = getMultiLayerConfiguration();
+
+        MultiLayerNetwork net = new MultiLayerNetwork(conf);
+        net.init();
+
+        net.setListeners(new ScoreIterationListener(20));
+
+        //Initialize the user interface backend
+        UIServer uiServer = UIServer.getInstance();
+
+        //Configure where the network information (gradients, score vs. time etc) is to be stored. Here: store in memory.
+        StatsStorage statsStorage = new InMemoryStatsStorage();         //Alternative: new FileStatsStorage(File), for saving and loading later
+
+        //Attach the StatsStorage instance to the UI: this allows the contents of the StatsStorage to be visualized
+        uiServer.attach(statsStorage);
+
+        //Then add the StatsListener to collect this information from the network, as it trains
+        net.setListeners(new StatsListener(statsStorage));
+
+        // ----- Train the network, evaluating the test set performance at each epoch -----
+
+        for (int i = 0; i < N_EPOCHS; i++) {
+            net.fit(trainData);
+            LOGGER.info("Epoch " + i + " complete. Time series evaluation:");
+
+            //Run regression evaluation on our single column input
+            RegressionEvaluation evaluation = new RegressionEvaluation(1);
+            INDArray features = testData.getFeatureMatrix();
+
+            INDArray lables = testData.getLabels();
+            INDArray predicted = net.output(features, false);
+
+            evaluation.evalTimeSeries(lables, predicted);
+
+            //Just do sout here since the logger will shift the shift the columns of the stats
+            System.out.println(evaluation.stats());
+        }
+
+        return net;
+
+    }
+
+    protected MultiLayerConfiguration getMultiLayerConfiguration() {
+        return new NeuralNetConfiguration.Builder()
+                .seed(140)
+                .optimizationAlgo(OptimizationAlgorithm.STOCHASTIC_GRADIENT_DESCENT)
+                .iterations(1)
+                .weightInit(WeightInit.XAVIER)
+                .updater(Updater.NESTEROVS).momentum(0.9)
+                .learningRate(LEARNING_RATE)
+                //.learningRate(0.00005) for train_date = 1000 ; test_data = 200
+                .list()
+                .layer(0, new GravesLSTM.Builder().activation(Activation.TANH).nIn(1).nOut(10)
+                        .build())
+                .layer(1, new RnnOutputLayer.Builder(LossFunctions.LossFunction.MSE)
+                        .activation(Activation.IDENTITY).nIn(10).nOut(1).build())
+                .build();
+    }
+
+}
