+++ anomaly-detection-train/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+build.version=@project.version@
