+++ anomaly-detection-train/src/main/resources/internet-traffic-data-in-bits-fr-timestamp-train.csv
