+++ anomaly-detection-train/src/test/resources/internet-traffic-data-in-bits-fr.csv
