--- src/main/java/com/example/AnomalyDetectionApplication.java
@@ -1,12 +1,0 @@
-package com.example;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AnomalyDetectionApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AnomalyDetectionApplication.class, args);
-	}
-}
