--- src/test/java/com/example/AnomalyDetectionTest.java
@@ -1,240 +1,0 @@
-package com.example;
-
-import org.datavec.api.records.reader.SequenceRecordReader;
-import org.datavec.api.records.reader.impl.csv.CSVSequenceRecordReader;
-import org.datavec.api.split.NumberedFileInputSplit;
-import org.datavec.api.util.ClassPathResource;
-import org.deeplearning4j.api.storage.StatsStorage;
-import org.deeplearning4j.datasets.datavec.SequenceRecordReaderDataSetIterator;
-import org.deeplearning4j.eval.RegressionEvaluation;
-import org.deeplearning4j.nn.api.OptimizationAlgorithm;
-import org.deeplearning4j.nn.conf.MultiLayerConfiguration;
-import org.deeplearning4j.nn.conf.NeuralNetConfiguration;
-import org.deeplearning4j.nn.conf.Updater;
-import org.deeplearning4j.nn.conf.layers.GravesLSTM;
-import org.deeplearning4j.nn.conf.layers.RnnOutputLayer;
-import org.deeplearning4j.nn.multilayer.MultiLayerNetwork;
-import org.deeplearning4j.nn.weights.WeightInit;
-import org.deeplearning4j.optimize.listeners.ScoreIterationListener;
-import org.deeplearning4j.ui.api.UIServer;
-import org.deeplearning4j.ui.stats.StatsListener;
-import org.deeplearning4j.ui.storage.InMemoryStatsStorage;
-import org.jfree.chart.ChartFactory;
-import org.jfree.chart.ChartPanel;
-import org.jfree.chart.JFreeChart;
-import org.jfree.chart.axis.NumberAxis;
-import org.jfree.chart.plot.PlotOrientation;
-import org.jfree.chart.plot.XYPlot;
-import org.jfree.chart.renderer.xy.XYDifferenceRenderer;
-import org.jfree.data.xy.XYSeries;
-import org.jfree.data.xy.XYSeriesCollection;
-import org.jfree.ui.RefineryUtilities;
-import org.junit.AfterClass;
-import org.junit.Test;
-import org.nd4j.linalg.activations.Activation;
-import org.nd4j.linalg.api.ndarray.INDArray;
-import org.nd4j.linalg.api.ops.impl.accum.MatchCondition;
-import org.nd4j.linalg.dataset.DataSet;
-import org.nd4j.linalg.dataset.api.iterator.DataSetIterator;
-import org.nd4j.linalg.dataset.api.preprocessor.NormalizerMinMaxScaler;
-import org.nd4j.linalg.factory.Nd4j;
-import org.nd4j.linalg.indexing.conditions.Conditions;
-import org.nd4j.linalg.lossfunctions.LossFunctions;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.swing.*;
-import java.awt.*;
-import java.io.File;
-import java.util.concurrent.CountDownLatch;
-
-/**
- * This example was inspired by SingleTimestepRegressionExample of deeplearning4j examples, found here:
-  https://github.com/deeplearning4j/dl4j-examples/blob/master/dl4j-examples/src/main/java/org/deeplearning4j/examples/recurrent/regression/SingleTimestepRegressionExample.java
- *
- * It demonstrates single time step regression using LSTM
- */
-public class AnomalyDetectionTest {
-    private static final Logger LOGGER = LoggerFactory.getLogger(AnomalyDetectionTest.class);
-
-    @Test
-    public void predict() throws Exception {
-
-        File baseDir = new ClassPathResource("/").getFile();
-        int miniBatchSize = 30;
-        //int miniBatchSize = 1000;
-
-        // ----- Load the training data -----
-        SequenceRecordReader trainReader = new CSVSequenceRecordReader(0, ";");
-        trainReader.initialize(new NumberedFileInputSplit(baseDir.getAbsolutePath() + "/internet-traffic-data-in-bits-fr-timestamp-train_%d.csv", 0, 0));
-
-        //For regression, numPossibleLabels is not used. Setting it to -1 here
-        DataSetIterator trainIter = new SequenceRecordReaderDataSetIterator(trainReader, miniBatchSize, -1, 1, true);
-
-        SequenceRecordReader testReader = new CSVSequenceRecordReader(0, ";");
-        testReader.initialize(new NumberedFileInputSplit(baseDir.getAbsolutePath() + "/internet-traffic-data-in-bits-fr-timestamp-test_%d.csv", 0, 0));
-        DataSetIterator testIter = new SequenceRecordReaderDataSetIterator(testReader, miniBatchSize, -1, 1, true);
-
-        //Create data set from iterator here since we only have a single data set
-        DataSet trainData = trainIter.next();
-        DataSet testData = testIter.next();
-
-        //Normalize data, including labels (fitLabel=true)
-        NormalizerMinMaxScaler normalizer = new NormalizerMinMaxScaler(0, 1);
-        normalizer.fitLabel(true);
-        normalizer.fit(trainData);              //Collect training data statistics
-
-        normalizer.transform(trainData);
-        normalizer.transform(testData);
-
-        // ----- Configure the network -----
-        MultiLayerConfiguration conf = new NeuralNetConfiguration.Builder()
-                .seed(140)
-                .optimizationAlgo(OptimizationAlgorithm.STOCHASTIC_GRADIENT_DESCENT)
-                .iterations(1)
-                .weightInit(WeightInit.XAVIER)
-                .updater(Updater.NESTEROVS).momentum(0.9)
-                .learningRate(0.00005)
-                //.learningRate(0.00005) for train_date = 1000 ; test_data = 200
-                .list()
-                .layer(0, new GravesLSTM.Builder().activation(Activation.TANH).nIn(1).nOut(10)
-                        .build())
-                .layer(1, new RnnOutputLayer.Builder(LossFunctions.LossFunction.MSE)
-                        .activation(Activation.IDENTITY).nIn(10).nOut(1).build())
-                .build();
-
-        MultiLayerNetwork net = new MultiLayerNetwork(conf);
-        net.init();
-
-        net.setListeners(new ScoreIterationListener(20));
-
-        //Initialize the user interface backend
-        UIServer uiServer = UIServer.getInstance();
-
-        //Configure where the network information (gradients, score vs. time etc) is to be stored. Here: store in memory.
-        StatsStorage statsStorage = new InMemoryStatsStorage();         //Alternative: new FileStatsStorage(File), for saving and loading later
-
-        //Attach the StatsStorage instance to the UI: this allows the contents of the StatsStorage to be visualized
-        uiServer.attach(statsStorage);
-
-        //Then add the StatsListener to collect this information from the network, as it trains
-        net.setListeners(new StatsListener(statsStorage));
-
-        // ----- Train the network, evaluating the test set performance at each epoch -----
-        int nEpochs = 200;
-        //int nEpochs = 2;
-
-        for (int i = 0; i < nEpochs; i++) {
-            net.fit(trainData);
-            LOGGER.info("Epoch " + i + " complete. Time series evaluation:");
-
-            //Run regression evaluation on our single column input
-            RegressionEvaluation evaluation = new RegressionEvaluation(1);
-            INDArray features = testData.getFeatureMatrix();
-
-            INDArray lables = testData.getLabels();
-            INDArray predicted = net.output(features, false);
-
-            evaluation.evalTimeSeries(lables, predicted);
-
-            //Just do sout here since the logger will shift the shift the columns of the stats
-            System.out.println(evaluation.stats());
-        }
-
-        //Init rrnTimeStemp with train data and predict test data
-        net.rnnTimeStep(trainData.getFeatureMatrix());
-        INDArray predicted_test = net.rnnTimeStep(testData.getFeatureMatrix());
-        INDArray predicted_train = net.rnnTimeStep(trainData.getFeatureMatrix());
-
-        //Revert data back to original values for plotting
-        normalizer.revert(trainData);
-        normalizer.revert(testData);
-        normalizer.revertLabels(predicted_test);
-        normalizer.revertLabels(predicted_train);
-
-        //INDArray error = predictions.sub(labels);
-        //INDArray anomalies_1 = testData.getLabels().gt(predicted_test);
-        INDArray anomalies = testData.getLabels().sub(predicted_test);
-        //One simple task is to count the number of values that match the condition
-
-        //ConditionBuilder builder = new ConditionBuilder();
-        //builder.and(Conditions.greaterThan(0.0)).and()
-        final double threshold = 100.0;
-        MatchCondition op = new MatchCondition(anomalies, Conditions.greaterThan(threshold));
-        int countGreaterThanZero = Nd4j.getExecutioner().exec(op,Integer.MAX_VALUE).getInt(0);  //MAX_VALUE = "along all dimensions" or equivalently "for entire array"
-        System.out.println("Number of values predicted values greater than actual values on test dataset: " + countGreaterThanZero+"/" + testData.getFeatures().data().length());
-
-
-        //Create plot with out data
-        XYSeriesCollection c = new XYSeriesCollection();
-        //createSeries(c, trainData.getFeatures(), 0, "Train data");
-        createSeries(c, testData.getFeatures(), 0, "Actual test data");
-        //createSeries(c, predicted_train, 0, "Predicted train data");
-        createSeries(c, predicted_test, 0, "Predicted test data");
-
-        plotDataset(c);
-
-        LOGGER.info("----- Example Complete -----");
-    }
-
-    private static XYSeriesCollection createSeries(XYSeriesCollection seriesCollection, INDArray data, int offset, String name) {
-        int nRows = data.shape()[2];
-        XYSeries series = new XYSeries(name);
-        for (int i = 0; i < nRows; i++) {
-            series.add(i + offset, data.getDouble(i));
-        }
-
-        seriesCollection.addSeries(series);
-
-        return seriesCollection;
-    }
-
-    /**
-     * Generate an xy plot of the datasets provided.
-     */
-    private static void plotDataset(XYSeriesCollection c) {
-
-        String title = "Regression example";
-        String xAxisLabel = "timestamp";
-        String yAxisLabel = "traffic data in bits";
-        PlotOrientation orientation = PlotOrientation.VERTICAL;
-        boolean legend = true;
-        boolean tooltips = true;
-        boolean urls = false;
-        JFreeChart chart = ChartFactory.createXYLineChart(title, xAxisLabel, yAxisLabel, c, orientation, legend, tooltips, urls);
-
-        // get a reference to the plot for further customisation...
-        final XYPlot plot = chart.getXYPlot();
-        plot.setDomainPannable(true);
-        XYDifferenceRenderer r = new XYDifferenceRenderer(Color.green,
-                Color.yellow, false);
-        r.setRoundXCoordinates(true);
-        plot.setDomainCrosshairLockedOnData(true);
-        plot.setRangeCrosshairLockedOnData(true);
-        plot.setDomainCrosshairVisible(true);
-        plot.setRangeCrosshairVisible(true);
-        plot.setRenderer(r);
-
-        // Auto zoom to fit time series in initial window
-        final NumberAxis rangeAxis = (NumberAxis) plot.getRangeAxis();
-        rangeAxis.setAutoRange(true);
-
-        JPanel panel = new ChartPanel(chart);
-
-        JFrame f = new JFrame();
-        f.add(panel);
-        f.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
-        f.pack();
-        f.setTitle("Training Data");
-
-        RefineryUtilities.centerFrameOnScreen(f);
-        f.setVisible(true);
-    }
-
-    @AfterClass
-    public static void tearDownClass() throws Exception {
-        //ugly : prevent test from ending so that we can have a look to chart
-        //CountDownLatch countdown = new CountDownLatch(1);
-        //countdown.await();
-    }
-}
